Former-commit-id: e4ee3ec2318723c58c963a23f0ae0f7687726aec

+++ i3/mkenv.sh
@@ -1,5 +1,26 @@
 #!/bin/sh
 
+. $HOME/myConfigs/gfw/get_apt_proxy.sh
+
+# install i3wm if not exist
+APT_SOURCE=$(grep debian.sur5r.net /etc/apt/sources.list)
+if [ -z "$APT_SOURCE" ]; then
+  echo "Adding i3wm official repository to '/etc/apt/sources.list'..."
+  echo "deb http://debian.sur5r.net/i3/ $(lsb_release -c -s) universe" | sudo tee --append /etc/apt/sources.list
+  echo "Update source..."
+  sudo apt-get $APT_PROXY update
+  echo "Install i3wm official repository key..."
+  sudo apt-get $APT_PROXY --allow-unauthenticated install sur5r-keyring
+  sudo apt-get $APT_PROXY update
+fi
+I3_PACKAGE=$(dpkg -l|cut -d " " -f 3|grep "^i3$")
+if [ -z "$I3_PACKAGE" ]; then
+  echo "Install i3wm..."
+  sudo apt-get $APT_PROXY install i3
+  echo "Install i3blocks..."
+  sudo apt-get $APT_PROXY install i3blocks
+fi
+
 CONFIG_HOME=$HOME/myConfigs/i3
 
 I3_HOME=$HOME/.i3
@@ -13,19 +34,59 @@ ln -sfnv $CONFIG_HOME/dunst/dunstrc $DUNST_HOME/dunstrc
 
 mkdir -p $HOME/bin
 ln -sfnv $CONFIG_HOME/i3bang/i3bang.rb $HOME/bin/i3bang
+# link default theme 'jellybeans' to ~/.i3/_config.colors
+ln -sfnv $CONFIG_HOME/colors/_config.jellybeans $I3_HOME/_config.colors
+# check if 'ruby' is installed or not
+RUBY_PACKAGE=$(dpkg -l|cut -d " " -f 3|grep "^ruby$")
+if [ -z "$RUBY_PACKAGE" ]; then
+  echo "Installing ruby..."
+  sudo apt-get $APT_PROXY install ruby
+  if [ "$?" -ne 0 ]; then
+    echo 'Install ruby failed, please check the output of apt-get.'
+    exit 1
+  fi
+  echo 'Installing ruby ... done'
+fi
 i3bang
 
 # check if 'consolekit' is installed or not
-PACKAGE=$(dpkg -l | grep consolekit)
-if [ -z "$PACKAGE" ]; then
+echo 'Checking package consolekit...'
+dpkg-query -l consolekit 2>/dev/null
+if [ "$?" -eq 1 ]; then
   # Install 'consolekit'
-  echo 'Install consolekit...'
-  sudo apt-get install consolekit
-  if [ "$?" -ne 0  ]; then
+  echo 'Installing consolekit...'
+  sudo apt-get $APT_PROXY install consolekit
+  if [ "$?" -ne 0 ]; then
     echo 'Install consolekit failed, please check the output of apt-get.'
     exit 1
   fi
-  echo 'Install consolekit ... done'
+  echo 'Installing consolekit ... done'
 fi
 
-sudo cp $CONFIG_HOME/xsessions/i3.desktop /usr/share/xsessions/i3.desktop
+if [ ! -e /usr/share/xsessions/i3.desktop ]; then
+  sudo cp $CONFIG_HOME/xsessions/i3.desktop /usr/share/xsessions/i3.desktop
+fi
+
+# xsession autostart files
+mkdir -p $HOME/.config/autostart
+_files="$CONFIG_HOME/xsessions/autostart/*.desktop"
+for file in $_files
+do
+  _name=`basename $file`
+  ln -sfnv $file $HOME/.config/autostart/$_name
+done
+
+# check if 'dex' is installed or not, it's needed to load xsession files
+echo 'Checking package dex...'
+dpkg-query -l dex 2>/dev/null
+if [ "$?" -eq 1 ]; then
+  # Install 'dex'
+  echo 'Installing dex...'
+  sudo apt-get $APT_PROXY install dex
+  if [ "$?" -ne 0 ]; then
+    echo 'Install dex failed, please check the output of apt-get.'
+    exit 1
+  fi
+  echo 'Installing dex ... done'
+fi
+
